fix: dynamic component plugin options
